meeting-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meeting-pipeline"
version = "0.1.0"
description = "Local batch meeting transcription and diarisation pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0",
  "rich>=13.7",
  "typer>=0.12",
  "watchdog>=4.0",
]

[project.optional-dependencies]
ml = [
  "faster-whisper>=1.0.3",
  "pyannote.audio>=3.3",
  "torch>=2.2",
  "torchaudio>=2.2",
]
dev = [
  "pytest>=8.1",
]

[project.scripts]
meetpipe = "app.cli:app"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
